+++ src/plugin.cpp
@@ -19,6 +19,7 @@ void init(rack::Plugin *p) {
 	p->addModel(modelMS);
 	p->addModel(modelDUKE);
 	p->addModel(modelMOIRE);
+	p->addModel(modelPILOT);
 	p->addModel(modelHUITRE);
 	p->addModel(modelOUAIVE);
 	p->addModel(modelPOUPRE);